# Faster bitwise-equal weld() implementation