Provides an alternative `weld()` implementation based on [Meshoptimizer](https://meshoptimizer.org/)'s approach. When tolerance=0, we use a hashtable to find bitwise-equal vertices quickly. This is vastly faster than previous approaches, but comes without tolerance options. When tolerance is non-zero, `weld()` falls back on the older implementation. In v4.0, the new implementation will be the default.

### Results

| model | before | after |
|-------|--------|-------|
| lovecraftian | 1,232ms | 149ms |
| plant_3 | 2,607ms | 431ms |
| concessionaire | 138ms | 32ms |
| weld_problem | 115ms | 21ms |

### Context

- Fixes https://github.com/donmccurdy/glTF-Transform/issues/738
- Related https://github.com/donmccurdy/glTF-Transform/pull/803
- Related https://github.com/donmccurdy/glTF-Transform/issues/1133

### Tasks

- [x] ~~debug performance regression above. hash collisions maybe?~~
- [x] add benchmark